Move the quick links functionality into separate components with some basic styles.
