+++ client/homescreen/layout.js
@@ -22,7 +22,6 @@ import {
 /**
  * Internal dependencies
  */
-import { QuickLinks } from '../quick-links';
 import StatsOverview from './stats-overview';
 import TaskListPlaceholder from '../task-list/placeholder';
 import InboxPanel from '../inbox-panel';
@@ -30,6 +29,7 @@ import { WelcomeModal } from './welcome-modal';
 
 import './style.scss';
 import '../dashboard/style.scss';
+import { StoreManagementLinks } from '../store-management-links';
 
 const TaskList = lazy( () =>
 	import( /* webpackChunkName: "task-list" */ '../task-list' )
@@ -93,7 +93,7 @@ export const Layout = ( {
 				>
 					{ isTaskListEnabled && renderTaskList() }
 					<StatsOverview />
-					{ ! isTaskListEnabled && <QuickLinks /> }
+					{ ! isTaskListEnabled && <StoreManagementLinks /> }
 				</div>
 			</Fragment>
 		);
