+++ client/store-management-links/quick-link-category/index.js
@@ -1,0 +1,20 @@
+/**
+ * External dependencies
+ */
+import React from '@wordpress/element';
+
+/**
+ * Internal dependencies
+ */
+import './style.scss';
+
+export const QuickLinkCategory = ( { title, children } ) => {
+	return (
+		<>
+			<h3 className="woocommerce-quick-links__category-header">
+				{ title }
+			</h3>
+			{ children }
+		</>
+	);
+};
