+++ client/store-management-links/quick-link-category/style.scss
@@ -1,0 +1,7 @@
+.woocommerce-quick-links__category-header {
+	margin: 8px 0;
+	text-transform: uppercase;
+	color: $gray-700;
+	line-height: 16px;
+	font-size: 11px;
+}
