+++ client/store-management-links/quick-links/index.js
@@ -1,0 +1,18 @@
+/**
+ * External dependencies
+ */
+import React from '@wordpress/element';
+
+/**
+ * Internal dependencies
+ */
+import './style.scss';
+
+export const QuickLinks = ( { children } ) => {
+	return <div className="woocommerce-quick-links">{ children }</div>;
+};
+
+//        <QuickLinks>
+//            <QuickLinkCategory title={"some title"}>
+//              <QuickLink />
+//              <QuickLink />
