+++ client/store-management-links/quick-links/style.scss
@@ -1,0 +1,2 @@
+.woocommerce-quick-links {
+}
