+++ client/store-management-links/style.scss
